requirements.txt
streamlit>=1.30.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
endee>=0.1.0
Pillow>=10.0.0
rapidocr-onnxruntime>=1.4.0
